worker/requirements.txt
# OCR Worker Dependencies
celery>=5.4.0
redis>=5.0.0
surya-ocr>=0.4.0
pdf2image>=1.17.0
Pillow>=10.0.0
opencv-python-headless>=4.10.0
pydantic>=2.10.0
supabase>=2.10.0
httpx>=0.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-image>=0.24.0
pytesseract>=0.3.10
